Move AssertMessageEqual to go1.7 buildflag file (#3050)
